+++ src/paths/segments/bspline.jl
@@ -299,7 +299,7 @@ end
 
 # positive curvature radius is a left handed turn, negative right handed.
 function curvatureradius(b::BSpline{T}, s) where {T}
-    t = arclength_to_t(b, s)
+    t = clamp(arclength_to_t(b, s), 0.0, 1.0)
     g = Interpolations.gradient(b.r, t)[1]
     h = Interpolations.hessian(b.r, t)[1]
     return (g[1]^2 + g[2]^2)^(3 // 2) / (g[1] * h[2] - g[2] * h[1])