+++ src/utils.jl
@@ -190,28 +190,61 @@ function discretize_curve(f, ddf, tolerance)
     return f.(ts)
 end
 
-function discretization_grid(ddf, tolerance)
-    bnds = (0.0, 1.0)
+function discretize_curve(s::Paths.Segment, tolerance)
+    return s.(discretization_grid(s, tolerance) * pathlength(s))
+end
 
+function discretize_curve(s::Paths.BSpline, tolerance)
+    return s.r.(discretization_grid(s, tolerance))
+end
+
+function discretization_grid(s::Paths.Segment, tolerance)
+    l = pathlength(s)
+    return discretization_grid(t -> Paths.signed_curvature(s, t * l), tolerance; t_scale=l)
+end
+
+function discretization_grid(s::Paths.BSpline, tolerance)
+    # Assume ds/dt ≈ 1 to avoid converting between arclength and t
+    h(t) = Paths.Interpolations.hessian(s.r, t)[1]
+    return discretization_grid(h, tolerance)
+end
+
+# Discretize using marching algorithm based on Hessian or curvature
+function discretization_grid(
+    ddf,
+    tolerance,
+    bnds::Tuple{Float64, Float64}=(0.0, 1.0);
+    t_scale=1.0
+)
     dt = 0.01
-    ts = zeros(typeof(bnds[1]), 4000)
+    ts = zeros(4000)
     ts[1] = bnds[1]
     t = bnds[1]
     i = 1
+    cc = norm(ddf(t))
     while t < bnds[2]
         i = i + 1
         i > length(ts) && error(
             "Too many points in curve for GDS. Increase discretization tolerance or split the curve."
         )
         t = ts[i - 1]
-        cc = norm(ddf(t))
-        if cc >= 1e-9 * oneunit(typeof(cc))
-            dt = sqrt(8 * tolerance / cc) #  Also assumes third derivative is small
+        # Set dt based on distance from chord assuming constant curvature
+        if cc >= 1e-9 * oneunit(typeof(cc)) # Update dt if curvature is not near zero
+            dt = uconvert(NoUnits, sqrt(8 * tolerance / cc) / t_scale)
         end
         if t + dt >= bnds[2]
             dt = bnds[2] - t
         end
+        # Check that curvature didn't increase too much (decrease is fine)
+        # Rare but may happen near inflection points
+        cc_next = norm(ddf(t + dt))
+        if (t_scale * dt)^2 * (cc_next - cc) / 24 > tolerance
+            dt = uconvert(NoUnits, sqrt(8 * tolerance / cc_next) / t_scale)
+            cc_next = norm(ddf(t + dt))
+        end
+        cc = cc_next
         ts[i] = min(bnds[2], t + dt)
+        t = ts[i]
     end
 
     return ts[1:i]