+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The format of this changelog is based on
 [Keep a Changelog](https://keepachangelog.com/), and this project adheres to
 [Semantic Versioning](https://semver.org/).
 
+## Upcoming
+
+  - Improved B-spline approximations
+
+### Fixed
+
+  - `atol` is now used for B-spline approximation of arbitrary curves when rendering to `SolidModel`
+
 ## 1.3.0 (2025-06-06)
 
   - Added `set_periodic!` to `SolidModels` to enable periodic meshes